begin edit/save flow
