+++ src/app-pages/reporting/reporting.scss
@@ -1,3 +1,8 @@
+.page-body {
+  margin-top: 6rem;
+  position: relative;
+}
+
 .right-panel {
   display: flex;
   justify-content: space-between;
