+++ src/css/index.scss
@@ -3,7 +3,6 @@ $success-color: #43ac6a;
 $warning-color: #d08002;
 $error-color: #f04124;
 
-
 .noselect {
   -webkit-touch-callout: none; /* iOS Safari */
   -webkit-user-select: none; /* Safari */
